Move chain payload builders to payload_builders:: module
