+++ crates/cosmos/cosmos-client-components/src/impls/client/create_client_payload.rs
@@ -1,6 +1,6 @@
 use cgp_core::CanRaiseError;
 use eyre::eyre;
-use hermes_relayer_components::chain::traits::components::create_client_payload_builder::CreateClientPayloadBuilder;
+use hermes_relayer_components::chain::traits::payload_builders::create_client::CreateClientPayloadBuilder;
 use hermes_relayer_components::chain::traits::types::create_client::{
     HasCreateClientOptionsType, HasCreateClientPayloadType,
 };
