+++ crates/mock/mock-cosmos-relayer/src/tests/ics02_client.rs
@@ -1,8 +1,8 @@
 use basecoin_store::impls::InMemoryStore;
-use hermes_relayer_components::chain::traits::components::create_client_payload_builder::CanBuildCreateClientPayload;
-use hermes_relayer_components::chain::traits::components::message_sender::CanSendMessages;
+use hermes_relayer_components::chain::traits::payload_builders::create_client::CanBuildCreateClientPayload;
 use hermes_relayer_components::chain::traits::queries::chain_status::CanQueryChainStatus;
 use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientState;
+use hermes_relayer_components::chain::traits::send_message::CanSendMessages;
 use hermes_relayer_components::relay::traits::components::update_client_message_builder::CanBuildUpdateClientMessage;
 use hermes_relayer_components::relay::traits::target::DestinationTarget;
 use hermes_relayer_components::runtime::traits::sleep::CanSleep;
