+++ crates/relayer/relayer-components-extra/src/components/extra/closures/chain/channel_handshake.rs
@@ -1,9 +1,9 @@
 use cgp_core::HasComponents;
-use hermes_relayer_components::chain::traits::components::channel_handshake_payload_builder::{
-    CanBuildChannelHandshakePayloads, ChannelHandshakePayloadBuilder,
-};
 use hermes_relayer_components::chain::traits::message_builders::channel_handshake::{
     CanBuildChannelHandshakeMessages, ChannelHandshakeMessageBuilder,
+};
+use hermes_relayer_components::chain::traits::payload_builders::channel_handshake::{
+    CanBuildChannelHandshakePayloads, ChannelHandshakePayloadBuilder,
 };
 use hermes_relayer_components::chain::traits::queries::chain_status::CanQueryChainHeight;
 use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientState;
