+++ crates/relayer/relayer-components/src/chain/traits/components/mod.rs
@@ -1,10 +1,2 @@
-pub mod ack_packet_payload_builder;
-pub mod channel_handshake_payload_builder;
-pub mod connection_handshake_payload_builder;
-pub mod create_client_payload_builder;
-pub mod message_sender;
 pub mod packet_fields_reader;
 pub mod packet_from_write_ack_builder;
-pub mod receive_packet_payload_builder;
-pub mod timeout_unordered_packet_payload_builder;
-pub mod update_client_payload_builder;
