+++ crates/relayer/relayer-components/src/chain/traits/mod.rs
@@ -2,5 +2,7 @@ pub mod components;
 pub mod event_subscription;
 pub mod logs;
 pub mod message_builders;
+pub mod payload_builders;
 pub mod queries;
+pub mod send_message;
 pub mod types;
