+++ crates/relayer/relayer-components/src/components/default/closures/relay/ack_packet_relayer.rs
@@ -1,15 +1,15 @@
 use cgp_core::{ErrorRaiser, HasComponents, HasErrorType};
 
-use crate::chain::traits::components::ack_packet_payload_builder::CanBuildAckPacketPayload;
-use crate::chain::traits::components::message_sender::CanSendMessages;
 use crate::chain::traits::components::packet_fields_reader::CanReadPacketFields;
-use crate::chain::traits::components::update_client_payload_builder::CanBuildUpdateClientPayload;
 use crate::chain::traits::message_builders::ack_packet::CanBuildAckPacketMessage;
 use crate::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
+use crate::chain::traits::payload_builders::ack_packet::CanBuildAckPacketPayload;
+use crate::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
 use crate::chain::traits::queries::chain_status::CanQueryChainStatus;
 use crate::chain::traits::queries::client_state::CanQueryClientState;
 use crate::chain::traits::queries::consensus_state::CanQueryConsensusState;
 use crate::chain::traits::queries::consensus_state_height::CanQueryConsensusStateHeight;
+use crate::chain::traits::send_message::CanSendMessages;
 use crate::chain::traits::types::chain_id::HasChainId;
 use crate::chain::traits::types::client_state::HasClientStateFields;
 use crate::chain::traits::types::consensus_state::HasConsensusStateType;
