+++ crates/relayer/relayer-components/src/relay/components/message_senders/chain_sender.rs
@@ -2,7 +2,7 @@ use alloc::vec::Vec;
 
 use cgp_core::async_trait;
 
-use crate::chain::traits::components::message_sender::CanSendMessages;
+use crate::chain::traits::send_message::CanSendMessages;
 use crate::chain::traits::types::ibc::HasIbcChainTypes;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::components::ibc_message_sender::IbcMessageSender;
