+++ crates/relayer/relayer-components/src/relay/components/packet_relayers/ack/base_ack_packet.rs
@@ -1,7 +1,7 @@
 use cgp_core::{async_trait, Async};
 
-use crate::chain::traits::components::ack_packet_payload_builder::CanBuildAckPacketPayload;
 use crate::chain::traits::message_builders::ack_packet::CanBuildAckPacketMessage;
+use crate::chain::traits::payload_builders::ack_packet::CanBuildAckPacketPayload;
 use crate::chain::traits::queries::client_state::CanQueryClientState;
 use crate::chain::traits::types::client_state::HasClientStateType;
 use crate::chain::traits::types::packet::HasIbcPacketTypes;
