+++ crates/relayer/relayer-components/src/relay/impls/channel/open_init.rs
@@ -1,8 +1,8 @@
 use cgp_core::async_trait;
 
-use crate::chain::traits::components::channel_handshake_payload_builder::CanBuildChannelHandshakePayloads;
-use crate::chain::traits::components::message_sender::CanSendSingleMessage;
 use crate::chain::traits::message_builders::channel_handshake::CanBuildChannelHandshakeMessages;
+use crate::chain::traits::payload_builders::channel_handshake::CanBuildChannelHandshakePayloads;
+use crate::chain::traits::send_message::CanSendSingleMessage;
 use crate::chain::traits::types::channel::HasInitChannelOptionsType;
 use crate::chain::traits::types::ibc_events::channel::HasChannelOpenInitEvent;
 use crate::relay::traits::chains::{CanRaiseRelayChainErrors, HasRelayChains};
