+++ crates/relayer/relayer-components/src/relay/impls/connection/open_init.rs
@@ -2,10 +2,10 @@ use core::iter::Iterator;
 
 use cgp_core::async_trait;
 
-use crate::chain::traits::components::connection_handshake_payload_builder::CanBuildConnectionHandshakePayloads;
-use crate::chain::traits::components::message_sender::CanSendSingleMessage;
 use crate::chain::traits::message_builders::connection_handshake::CanBuildConnectionHandshakeMessages;
+use crate::chain::traits::payload_builders::connection_handshake::CanBuildConnectionHandshakePayloads;
 use crate::chain::traits::queries::client_state::CanQueryClientState;
+use crate::chain::traits::send_message::CanSendSingleMessage;
 use crate::chain::traits::types::connection::HasInitConnectionOptionsType;
 use crate::chain::traits::types::ibc_events::connection::HasConnectionOpenInitEvent;
 use crate::relay::traits::chains::{CanRaiseRelayChainErrors, HasRelayChains};
