+++ crates/relayer/relayer-components/src/relay/traits/components/update_client_message_builder.rs
@@ -2,7 +2,7 @@ use alloc::vec::Vec;
 
 use cgp_core::prelude::*;
 
-use crate::chain::traits::components::message_sender::CanSendMessages;
+use crate::chain::traits::send_message::CanSendMessages;
 use crate::chain::types::aliases::{HeightOf, MessageOf};
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::target::ChainTarget;
