+++ crates/solomachine/solomachine-relayer/src/impls/chain/solomachine_components/create_client_payload.rs
@@ -1,5 +1,5 @@
 use cgp_core::prelude::*;
-use hermes_relayer_components::chain::traits::components::create_client_payload_builder::CreateClientPayloadBuilder;
+use hermes_relayer_components::chain::traits::payload_builders::create_client::CreateClientPayloadBuilder;
 
 use crate::traits::solomachine::Solomachine;
 use crate::types::chain::SolomachineChain;
