+++ crates/solomachine/solomachine-relayer/src/impls/chain/solomachine_components/receive_packet_payload.rs
@@ -1,6 +1,6 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_client_components::methods::encode::encode_protobuf;
-use hermes_relayer_components::chain::traits::components::receive_packet_payload_builder::ReceivePacketPayloadBuilder;
+use hermes_relayer_components::chain::traits::payload_builders::receive_packet::ReceivePacketPayloadBuilder;
 use ibc_relayer_types::core::ics04_channel::packet::Packet;
 use ibc_relayer_types::core::ics24_host::path::CommitmentsPath;
 use ibc_relayer_types::Height;
